Refactor hdd code to allow for better concurrent work.
