+++ Src/inc/TestDefinition.h
@@ -10,7 +10,27 @@ extern void testFunction_MAG_Reads();
 extern void testFunction_LogTimer_Callback();
 extern void testerFunction_PWMGeneration();
 extern void testFunction_IMU_Dupe_Reads();
+extern int testFuncID_HDD();
 extern void testFunction_HDD();
+
+// if your id is not present, add it
+extern void testFunction_HDD1();
+extern void testFunction_HDD2();
+extern void testFunction_HDD3();
+extern void testFunction_HDD4();
+extern void testFunction_HDD5();
+extern void testFunction_HDD6();
+
+// add your id's test function below
+static void (*testFunctions_HDD[6])() = {
+		testFunction_HDD,
+		testFunction_HDD1,
+		testFunction_HDD2,
+		testFunction_HDD3,
+		testFunction_HDD4,
+		testFunction_HDD5,
+		testFunction_HDD6
+};
 
 //-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-
 //                           TEST REGISTRATION
@@ -33,7 +53,7 @@ void (*getTestFunction(int test_id))(void) {
 		case 8:
 			return testFunction_IMU_Dupe_Reads;
 		case 10:
-			return testFunction_HDD;
+			return testFunctions_HDD[testFuncID_HDD()];
 		default:
 			return testFunction_Nothing;
 	}
